fix: Fixed issues caused by cleaning up .NET objects. Keeping the JavaScript and .NET objects in sync causes more overhead and management that just trigging a delete during a .NET object finalize call.
Because the .NET object can go out of scope, for some reason or another, it could cause the JavaScript reference to be lost, when the cache is cleared.

To make this library as easy as possible to use without having to micro manage the state of all objects created/destroyed, it will just hold onto the cache for the life of the browser page request.

The automatic cache remove was reverted, the 'removeEntity' interop is still available for custom code to be easily created.

General Styles and Cleanup needed after adding .editorconfig.